+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 
 - `limestone/Card` audio guidance for disabled
 - `limestone/Card` `icon` prop to allow user to change the icon
+- `limestone/Card` `splitCaption` prop to allow user to split the content for the captions
 - `limestone/Card` added option for progressBar
 - `limestone/Icon` supported icon list, updating existing icons